requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-CORS>=4.0.0
Flask-Babel>=4.0.0
SQLAlchemy>=2.0.0
pymysql>=1.1.0
pandas>=2.0.0
python-dateutil>=2.8.0
Flask-Login>=0.6.3
Flask-Bcrypt>=1.0.1
psycopg2-binary>=2.9.0
waitress>=2.1.0
openpyxl>=3.1.0
waitress>=2.1.0
